+++ packages/windie-sdk-js/cjs/runtime/AgentClient.js
@@ -274,7 +274,7 @@ class AgentClient {
         if (explicitRuntime) {
             return explicitRuntime;
         }
-        const daemonOptions = this.defaultOptions.localRuntimeDaemon ?? this.defaultOptions.sidecarDaemon;
+        const daemonOptions = this.defaultOptions.localRuntimeDaemon;
         if (daemonOptions) {
             return new LocalSidecarRuntime_js_1.AgentLocalRuntimeHttpClient({
                 ...daemonOptions,