+++ packages/windie-sdk-js/cjs/runtime/WindieChatSession.js
@@ -18,6 +18,6 @@ var __exportStar = (this && this.__exportStar) || function(m, exports) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.WindieChatSession = void 0;
-const AgentChatSession_js_1 = require("./AgentChatSession.js");
 __exportStar(require("./AgentChatSession.js"), exports);
+var AgentChatSession_js_1 = require("./AgentChatSession.js");
 Object.defineProperty(exports, "WindieChatSession", { enumerable: true, get: function () { return AgentChatSession_js_1.AgentChatSession; } });