+++ packages/windie-sdk-js/cjs/stores/SidecarConversationStore.js
@@ -4,6 +4,6 @@
  */
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.SidecarConversationStore = exports.LocalRuntimeConversationStore = void 0;
-const LocalRuntimeConversationStore_js_1 = require("./LocalRuntimeConversationStore.js");
+var LocalRuntimeConversationStore_js_1 = require("./LocalRuntimeConversationStore.js");
 Object.defineProperty(exports, "LocalRuntimeConversationStore", { enumerable: true, get: function () { return LocalRuntimeConversationStore_js_1.LocalRuntimeConversationStore; } });
 Object.defineProperty(exports, "SidecarConversationStore", { enumerable: true, get: function () { return LocalRuntimeConversationStore_js_1.LocalRuntimeConversationStore; } });